% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{add.cummulative.start.end}
\alias{add.cummulative.start.end}
\title{add cummulaative start / end columns too order the genome chr1 -> chrX}
\usage{
add.cummulative.start.end(input, add.GrCh37.scale = TRUE)
}
\description{
add cummulaative start / end columns too order the genome chr1 -> chrX
}
