% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{cal.Subject.accross.regions}
\alias{cal.Subject.accross.regions}
\title{Function to adjust any quantification (subject) for cellularity (cell) and then average accross regions to get 1 quantification per tumour}
\usage{
cal.Subject.accross.regions(SubjectbyRegion, CellbyRegion)
}
\description{
Function to adjust any quantification (subject) for cellularity (cell) and then average accross regions to get 1 quantification per tumour
}
