% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{calc.ploidy}
\alias{calc.ploidy}
\title{#Function to calculate ploidy distribution and main ploidy per sample from seg input}
\usage{
calc.ploidy(seg, cnCol = 6, check.names = FALSE)
}
\description{
#Function to calculate ploidy distribution and main ploidy per sample from seg input
}
