% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{calc.wgii}
\alias{calc.wgii}
\title{Function to caulcated wgII from seg CN data}
\usage{
calc.wgii(seg, threshold = 0.6, check.names = FALSE, include.sexchrom = FALSE)
}
\description{
Function to caulcated wgII from seg CN data
}
