% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{correct.specific.ps}
\alias{correct.specific.ps}
\title{Function to FDR correct a set of p-values even when duplicates exist and only a subset is being considered for correction}
\usage{
correct.specific.ps(
  table,
  field.to.deduplicate = NA,
  field.to.correct,
  indices.to.correct = "all",
  method = "BH"
)
}
\description{
Function to FDR correct a set of p-values even when duplicates exist and only a subset is being considered for correction
}
