% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{extract.tree.mutTable}
\alias{extract.tree.mutTable}
\title{function to exact tree that hase been overlaid onto a TRACERx mutTable}
\usage{
extract.tree.mutTable(mutTable, clone.field, patient.field = NA, parent.field)
}
\description{
function to exact tree that hase been overlaid onto a TRACERx mutTable
}
