% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{extract_ccf_table}
\alias{extract_ccf_table}
\title{Function to extract CCF tables from a TRACERx mutTable containing many samples}
\usage{
extract_ccf_table(
  muttable_df,
  ccf_col = "PhyloCCF_MPhase",
  sampleid_col = "tumour_id",
  cluster_col = "PyCloneCluster_MPhase",
  driver_col = "DriverMut"
)
}
\description{
Function to extract CCF tables from a TRACERx mutTable containing many samples
}
