% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{integrate.mutTables}
\alias{integrate.mutTables}
\title{function to merge MutTables ensuring the sampleID formats etc are consistent}
\usage{
integrate.mutTables(mutTableAllpriority, mutTableAllextra)
}
\description{
function to merge MutTables ensuring the sampleID formats etc are consistent
}
