% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{interpret_following_Args}
\alias{interpret_following_Args}
\title{Function to allow R script to interpret following arguments specified in terminal with flags}
\usage{
interpret_following_Args(args, arg.table)
}
\description{
Function to allow R script to interpret following arguments specified in terminal with flags
}
