% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{make.axis.adjust.cummulative}
\alias{make.axis.adjust.cummulative}
\title{Function to make a cummulative x axis for chromosomes across the genome}
\usage{
make.axis.adjust.cummulative(input.with.cummulative = NA, chrlength = NA)
}
\description{
Function to make a cummulative x axis for chromosomes across the genome
}
