% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{Overlay.tree.parents}
\alias{Overlay.tree.parents}
\title{function to overlay tree information onto a TRACERx mutTable}
\usage{
Overlay.tree.parents(mutTable, trees, clone.field, patient.field)
}
\description{
function to overlay tree information onto a TRACERx mutTable
}
