% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{read_trees}
\alias{read_trees}
\title{Function to extract the tree information from all patients in a tree output directory}
\usage{
read_trees(tree_path, return_all = FALSE, remove_missing = TRUE)
}
\description{
Function to extract the tree information from all patients in a tree output directory
}
