% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{rename.TxIDs.in.mutid}
\alias{rename.TxIDs.in.mutid}
\title{function to rename TRACERx mutation ids with correct most up to datte patient ids}
\usage{
rename.TxIDs.in.mutid(mutids)
}
\description{
function to rename TRACERx mutation ids with correct most up to datte patient ids
}
