% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankellA.functions.R
\name{trx_rename.fn}
\alias{trx_rename.fn}
\title{Function to rename any TRACERx patient ids to the latest version}
\usage{
trx_rename.fn(trxid, trialID = "LTX")
}
\description{
Function to rename any TRACERx patient ids to the latest version
}
